src/test/java/com/inowhite/cosmos/core/controller/ControllerUtilTest.java
/*
 * Cosmos - IT Management and Service Desk System
 * Copyright (C) 2020  Gabriel Ribeiro
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package com.inowhite.cosmos.core.controller;

import org.junit.jupiter.api.Test;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.Arrays;

import static org.junit.jupiter.api.Assertions.assertEquals;

@SpringBootTest
public class ControllerUtilTest {

  @Test
  public void testGetOrderSpecAsList() {
    final String specs = "+id;-createdAt";
    var result = ControllerUtil.getOrderSpecAsList(specs);
    var expected = Arrays.asList("+id", "-createdAt");
    assertEquals(expected.get(0), result.get(0));
    assertEquals(expected.get(1), result.get(1));
  }

}
